Moved read/write to MCU1, also only now use 2 data pins for comms.
Updated some code to account for this change. As well as the fact that the cs0 cs1 lines were reversed.
Updated some docs for new swapped cs line values

+++ Dreamcast/sw/shared/serial_bridge/serial_bridge.h
@@ -22,4 +22,4 @@ void interconnect_tx16(uint16_t value);
 
 bool interconnect_rx_buffer_has_data();
 bool interconnect_rx_has_data();
-uint16_t interconnect_rx_get(); // Read 16 bits of the rx buffer
+uint8_t interconnect_rx_get(); // Read 8 bits of the rx buffer
